requirements.txt
python-dotenv
flask
flask_sqlalchemy
flask_login
psycopg2
requests